refactoring date picker field ui tests
