+++ app/src/androidTest/java/br/com/rbrthmn/ui/financialcompanion/components/DatePickerFieldUITest.kt
@@ -1,0 +1,72 @@
+/*
+ *
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Modifications made by Roberto Kenzo Hamano, 2024
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package br.com.rbrthmn.ui.financialcompanion.components
+
+import androidx.activity.ComponentActivity
+import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.assertIsNotDisplayed
+import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.onNodeWithContentDescription
+import androidx.compose.ui.test.onNodeWithTag
+import androidx.compose.ui.test.performClick
+import br.com.rbrthmn.R
+import br.com.rbrthmn.ui.financialcompanion.BaseUITest
+import br.com.rbrthmn.ui.financialcompanion.onNodeWithStringId
+import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.CALENDAR_TEST_TAG
+import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.DatePickerField
+import org.junit.Test
+
+class DatePickerFieldUITest : BaseUITest() {
+    override val composeTestRule = createAndroidComposeRule<ComponentActivity>()
+
+    override fun setup() =
+        composeTestRule.setContent { DatePickerField(onDateSelected = {}, isError = false) }
+
+
+    @Test
+    fun calendar_icon_button_should_show_dialog() {
+        composeTestRule.run {
+            onNodeWithContentDescription(activity.getString(R.string.date_picker_title)).performClick()
+
+            onNodeWithTag(CALENDAR_TEST_TAG).assertIsDisplayed()
+        }
+    }
+
+    @Test
+    fun ok_button_should_dismiss_calendar_dialog() {
+        composeTestRule.run {
+            onNodeWithContentDescription(activity.getString(R.string.date_picker_title)).performClick()
+            onNodeWithStringId(R.string.date_picker_ok).performClick()
+
+            onNodeWithTag(CALENDAR_TEST_TAG).assertIsNotDisplayed()
+        }
+    }
+
+    @Test
+    fun cancel_button_should_dismiss_calendar_dialog() {
+        composeTestRule.run {
+            onNodeWithContentDescription(activity.getString(R.string.date_picker_title)).performClick()
+            onNodeWithStringId(R.string.date_picker_cancel).performClick()
+
+            onNodeWithTag(CALENDAR_TEST_TAG).assertIsNotDisplayed()
+        }
+    }
+}
